fix: implement various PR requests
