+++ packages/tests/utils/db.ts
@@ -1,10 +1,7 @@
 import { PrismaClient } from 'database';
 
 export const prisma = new PrismaClient({
-  log:
-    process.env.NODE_ENV === 'development'
-      ? ['query', 'error', 'warn']
-      : ['error'],
+  log: ['error'],
 });
 
 /**
