+++ packages/tests/utils/users.ts
@@ -3,27 +3,32 @@ import { prisma } from './db';
 
 interface ITestUser {
   userId: string;
+  email: string;
   role: Roles;
 }
 
-/* A list of dumby users with different roles */
+/* A list of dummy users with different roles */
 export const testUserBasic: ITestUser = {
   userId: 'clglqupom00bp5s87fxip254i',
+  email: 'basic@waldo.vision',
   role: Roles.USER,
 };
 
 export const testUserTrusted: ITestUser = {
   userId: 'clglqupom00bp5s87fxip254a',
+  email: 'trusted@waldo.vision',
   role: Roles.TRUSTED,
 };
 
 export const testUserMod: ITestUser = {
   userId: 'clglqupom00bp5s87fxip254b',
+  email: 'mod@waldo.vision',
   role: Roles.MOD,
 };
 
 export const testUserAdmin: ITestUser = {
   userId: 'clglqupom00bp5s87fxip254c',
+  email: 'admin@waldo.vision',
   role: Roles.ADMIN,
 };
 
@@ -40,7 +45,7 @@ export const allTestUsers = [
  * @returns a cookie string
  */
 export const createUserCookie = (user: ITestUser): string => {
-  return `userId=${user.userId}; provider=dumby; role=${user.role}`;
+  return `userId=${user.userId}; provider=dummy; role=${user.role}`;
 };
 
 /**
@@ -51,6 +56,7 @@ export const setUpUsers = async () => {
     data: allTestUsers.map(user => ({
       id: user.userId,
       name: user.userId,
+      email: user.email,
       role: user.role,
     })),
   });
